requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
Flask-JWT-Extended==3.24.1
